Use library gfoidl.Base64 library for Base64-URL encoding instead of custom implementation
