+++ perf/Benchmarks/Benchmarks.csproj
@@ -38,7 +38,7 @@
     <PackageReference Include="JWT" Version="7.2.1" />
     <PackageReference Include="Microsoft.IdentityModel.JsonWebTokens" Version="6.6.0" />
     <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.6.0" />
-    <PackageReference Include="gfoidl.Base64" Version="1.0.1" />
+    <PackageReference Include="gfoidl.Base64" Version="1.1.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\JsonWebToken\JsonWebToken.csproj" />
